# Server/swaggerconfig.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PSF: Portugal Food Sustainability",
        "description": "API - Desenvolvimento sustentavel",
        "termsOfService": "http://localhost:5000/terms",
        "contact": {
            "name": "Developers - PDS 2022 - LESI",
            "url": "https://www.ipca.pt",
            "email": "codesolutions@gmail.com"
        },
        "license": {
            "name": "Use under IPCA rights",
            "url": "https://www.ipca.pt/license"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/produto/publicarProduto": {
            "post": {
                "description": "Publicar um novo Produto",
                "tags": [
                    "Produtos"
                ],
                "summary": "publicar um produto",
                "operationId": "Produtos_publicarProduto",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Produto"
                            },
                            "examples": {
                                "Produto": {
                                    "value": {
                                        "name": "oi",
                                        "image": "alo",
                                        "quantity": 333,
                                        "price": 23,
                                        "hourRecoMin": "10/10/10",
                                        "hourRecoMax": "10/10/12",
                                        "lojaId": 1,
                                        "subCatProdId": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not found!"
                    },
                    "201": {
                        "description": "Created Post!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Produto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produto/editarProduto": {
            "post": {
                "description": "Editar um Produto",
                "tags": [
                    "Produtos"
                ],
                "summary": "editar um produto",
                "operationId": "Produtos_editarProduto",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EditarProduto"
                            },
                            "examples": {
                                "Produto": {
                                    "value": {
                                        "id": 1,
                                        "name": "oi",
                                        "image": "alo",
                                        "quantity": 333,
                                        "price": 23,
                                        "hourRecoMin": "10/10/10",
                                        "hourRecoMax": "10/10/12",
                                        "lojaId": 1,
                                        "subCatProdId": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Forbidden!"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "201": {
                        "description": "Created Post!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/EditarProduto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produto/listarProdutos": {
            "get": {
                "description": "Listar Produtos",
                "tags": [
                    "Produtos"
                ],
                "summary": "listar todos os produtos",
                "operationId": "Produtos_listarProduto",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden!"
                    },
                    "200": {
                        "description": "OK!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/ListarProduto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Efetuar Login",
                "tags": [
                    "Security"
                ],
                "summary": "efetuar o login",
                "operationId": "Security_login",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Login"
                            },
                            "examples": {
                                "Login": {
                                    "value": {
                                        "username": "silva.pastelaria@outlook.com",
                                        "password": "pastelaria1774"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not Found!"
                    },
                    "201": {
                        "description": "Created Post!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Login"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/registeruser": {
            "post": {
                "description": "Efetuar um Registo",
                "tags": [
                    "Security"
                ],
                "summary": "registar um utilizador",
                "operationId": "Security_registeruser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Registeruser"
                            },
                            "examples": {
                                "Cliente": {
                                    "value": {
                                        "name": "name example",
                                        "password": "123456789",
                                        "email": "name@example.pt",
                                        "contact": "965321456",
                                        "nif": "123456789",
                                        "permission": "3"
                                    }
                                },
                                "Loja": {
                                    "value": {
                                        "name": "name example",
                                        "password": "123456789",
                                        "email": "name@example.pt",
                                        "contact": "name@example.pt",
                                        "nif": "123456789",
                                        "permission": "2"
                                    }
                                },
                                "Admin": {
                                    "value": {
                                        "name": "name example",
                                        "password": "123456789",
                                        "email": "name@example.pt",
                                        "contact": "name@example.pt",
                                        "nif": "123456789",
                                        "permission": "1"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "201": {
                        "description": "Created Post!"
                    }
                }
            }
        },
        "/mostarPerfil": {
            "post": {
                "description": "Mostrar Perfil de um Utilizador",
                "tags": [
                    "Perfil"
                ],
                "summary": "mostrar o perfil do utilizador",
                "operationId": "Security_mostarPerfil",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MostarPerfil"
                            },
                            "examples": {
                                "Utilizador": {
                                    "value": {
                                        "username": "silva.pastelaria@outlook.com"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not Found!"
                    },
                    "201": {
                        "description": "Created Post!"
                    }
                }
            }
        },
        "/candidaturaLoja": {
            "post": {
                "description": "Candidaturas das Lojas",
                "tags": [
                    "Candidaturas"
                ],
                "summary": "candidatura de uma loja",
                "operationId": "Candidaturas_candidatura",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "morada": {
                                        "type": "string"
                                    },
                                    "nif": {
                                        "type": "integer"
                                    },
                                    "adminloja": {
                                        "type": "integer"
                                    },
                                    "categoria": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not Found!"
                    },
                    "500": {
                        "description": "Internal Error!"
                    },
                    "201": {
                        "description": "Created Post!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Candidatura"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/candidaturaLojaFicheiro/{id}": {
            "post": {
                "description": "Candidaturas das Lojas com  Ficheiros",
                "tags": [
                    "Candidaturas"
                ],
                "summary": "enviar ficheiros para a candidatura de uma loja",
                "operationId": "Candidaturas_candidaturaLojaFicheiro",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "x-nullable": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not Found!"
                    },
                    "500": {
                        "description": "Internal Error!"
                    },
                    "201": {
                        "description": "Created Post!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/CandidaturaFicheiro"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Login": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
            "Produto": {
                "type": "object",
                "properties": {
                    "idProduto": {
                        "type": "integer"
                    },
                    "nomeProduto": {
                        "type": "string"
                    },
                    "quantidade": {
                        "type": "integer"
                    },
                    "preco": {
                        "type": "number"
                    },
                    "horaRecolhaMin": {
                        "type": "string"
                    },
                    "horaRecolhaMax": {
                        "type": "string"
                    },
                    "fotoProduto": {
                        "type": "string"
                    },
                    "lojaId": {
                        "type": "integer"
                    },
                    "SubcategoriaProdId": {
                        "type": "integer"
                    }
                }
            },
            "EditarProduto": {
                "type": "object",
                "properties": {
                    "idProduto": {
                        "type": "integer"
                    },
                    "nomeProduto": {
                        "type": "string"
                    },
                    "quantidade": {
                        "type": "integer"
                    },
                    "preco": {
                        "type": "number"
                    },
                    "horaRecolhaMin": {
                        "type": "string"
                    },
                    "horaRecolhaMax": {
                        "type": "string"
                    },
                    "fotoProduto": {
                        "type": "string"
                    },
                    "lojaId": {
                        "type": "integer"
                    },
                    "SubcategoriaProdId": {
                        "type": "integer"
                    }
                }
            },
            "ListarProduto": {
                "type": "object",
                "properties": {
                    "idProduto": {
                        "type": "integer"
                    },
                    "nomeProduto": {
                        "type": "string"
                    },
                    "quantidade": {
                        "type": "integer"
                    },
                    "preco": {
                        "type": "number"
                    },
                    "horaRecolhaMin": {
                        "type": "string"
                    },
                    "horaRecolhaMax": {
                        "type": "string"
                    },
                    "fotoProduto": {
                        "type": "string"
                    },
                    "lojaId": {
                        "type": "integer"
                    },
                    "SubcategoriaProdId": {
                        "type": "integer"
                    }
                }
            },
            "MostarPerfil": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "number"
                    },
                    "contacto": {
                        "type": "string"
                    }
                }
            },
            "CandidaturaFicheiro": {
                "type": "object",
                "properties": {
                    "documento": {
                        "type": "string"
                    },
                    "lojaid": {
                        "type": "integer"
                    }
                }
            },
            "Candidatura": {
                "type": "object",
                "properties": {
                    "idLoja": {
                        "type": "integer"
                    },
                    "morada": {
                        "type": "string"
                    },
                    "nif": {
                        "type": "integer"
                    },
                    "adminloja": {
                        "type": "integer"
                    },
                    "categoria": {
                        "type": "integer"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}